% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cases.R, R/enrich.R
\docType{package}
\name{AEenrich-package}
\alias{AEenrich-package}
\title{AEenrich: Adverse Event Enrichment Tests}
\description{
The count_cases function converts report-level data into aggregated data by
grouping on specified covariates. Use \code{count_cases} to summarize adverse
event reports into analysis-ready tables. See the GitHub home page at
\url{https://github.com/umich-biostatistics/AEenrich} or run \code{?count_cases}
for examples.

Perform adverse event enrichment tests. The \code{enrich} function implements
enrichment analysis for adverse event count data, which contain many zeros and
ties. Two methods are provided: \code{AEFisher}, a modified Fisher’s exact test
based on pre-selected significant adverse events, and \code{AEKS}, a modified
Kolmogorov–Smirnov statistic. See the GitHub home page at
\url{https://github.com/umich-biostatistics/AEenrich} or run \code{?enrich} for
examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umich-biostatistics/AEenrich}
  \item Report bugs at \url{https://github.com/umich-biostatistics/AEenrich/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/umich-biostatistics/AEenrich}
  \item Report bugs at \url{https://github.com/umich-biostatistics/AEenrich/issues}
}

}
\author{
\strong{Maintainer}: Michael Kleinsasser \email{mkleinsa@umich.edu}

Authors:
\itemize{
  \item Shuoran Li \email{shuoranl@umich.edu}
  \item Hongfan Chen \email{chenhf@umich.edu}
  \item Lili Zhao \email{zhaolili@med.umich.edu}
}

}
