% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid1}
\alias{covid1}
\title{Covid Vaccine Adverse Event Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12500 rows and 5 columns.
}
\usage{
covid1
}
\value{
A data frame with the following columns:
\item{VAERS_ID}{Event ID}
\item{VAX_LABEL}{Vaccine type}
\item{AE_NAME}{Adverse event name}
\item{AGE}{Covariate}
\item{SEX}{Covariate}
}
\description{
Adverse event data in the long format. Each row is a single adverse
event, along with covariates.
}
\keyword{datasets}
