% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid2}
\alias{covid2}
\title{Covid Vaccine Adverse Event Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2656 rows and 6 columns.
}
\usage{
covid2
}
\value{
A data frame with the following columns:
\item{DRUG_TYPE}{Vaccine type}
\item{AE_NAME}{Adverse event name}
\item{AEYes}{Number of observations that have this adverse event}
\item{AENo}{Number of observations that do not have this adverse event}
\item{AGE}{Covariate}
\item{SEX}{Covariate}
}
\description{
Adverse event data in the short format. Each row is a count of adverse
events with the given name.
}
\keyword{datasets}
