% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flu1}
\alias{flu1}
\title{Flu Vaccine Adverse Event Data}
\format{
An object of class \code{data.frame} with 12622 rows and 3 columns.
}
\usage{
flu1
}
\description{
Adverse event data in the long format. Each row is a single adverse
event.
}
\details{
\itemize{
\item VAERS_ID Event ID
\item VAX_TYPE Vaccine type
\item AE_NAME Adverse event name
}
}
\keyword{datasets}
