% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flu2}
\alias{flu2}
\title{Flu Vaccine Adverse Event Data}
\format{
An object of class \code{data.frame} with 1603 rows and 3 columns.
}
\usage{
flu2
}
\description{
Adverse event data in the short format. Each row is a count of adverse
events with the given name.
}
\details{
\itemize{
\item VAX_TYPE Vaccine type
\item AE_NAME Adverse event name
\item Count Frequency of adverse event
}
}
\keyword{datasets}
