% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group}
\alias{group}
\title{Group Structure Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35339 rows and 2 columns.
}
\usage{
group
}
\value{
A data frame with the following columns:
\item{AE_NAME}{Adverse event name}
\item{GROUP_NAME}{Group name}
}
\description{
Identifies which group each set of adverse events belongs.
}
\keyword{datasets}
