% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\docType{package}
\name{AEenrich-package}
\alias{AEenrich}
\alias{AEenrich-package}
\title{AEenrich: Adverse Event Enrichment Tests}
\description{
Gene enrichment tests to perform adverse event (AE) enrichment
analysis. Unlike the continuous gene expression data, AE data are counts.
Therefore, AE data has many zeros and ties. We propose two enrichment tests.
One is a modified Fisher's exact test based on pre-selected significant AEs,
while the other is based on a modified Kolmogorov-Smirnov statistic.

Use the function \code{enrich} to fit models and inspect results.

See our \href{https://github.com/umich-biostatistics/AEenrich}{Github home page}
or run ?enrich for examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umich-biostatistics/AEenrich}
  \item Report bugs at \url{https://github.com/umich-biostatistics/AEenrich/issues}
}

}
\author{
\strong{Maintainer}: Michael Kleinsasser \email{mkleinsa@umich.edu}

Authors:
\itemize{
  \item Shuoran Li \email{shuoranl@umich.edu}
  \item Lili Zhao \email{zhaolili@med.umich.edu}
}

}
