% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group}
\alias{group}
\title{Group Structure Data}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
group
}
\description{
Identifies which group each set of adverse events belongs.
}
\details{
\itemize{
\item AE_NAME Adverse event name
\item GROUP_NAME Group name
}
}
\keyword{datasets}
